Changed file handling, hopefully fixes bad single line indel at the beginning of the indel writes in the simulation (speculate that is may be file io load)

+++ prob2020/python/indel.py
@@ -139,7 +139,7 @@ def compute_indel_length(fs_df):
     indel_len = pd.Series(index=fs_df.index)
     indel_len[fs_df['Reference_Allele']=='-'] = fs_df['Tumor_Allele'][fs_df['Reference_Allele']=='-'].str.len()
     indel_len[fs_df['Tumor_Allele']=='-'] = fs_df['Reference_Allele'][fs_df['Tumor_Allele']=='-'].str.len()
-    indel_len = indel_len.fillna(0)
+    indel_len = indel_len.fillna(0).astype(int)
     return indel_len
